// src/types/Coin.ts
export interface Coin {
  id: string;
  name: string;
  symbol: string;
  exchangeRate: number;
  liquidity: number;
  rank: number;
  rankTrend: string;
  utility: string;
  marketCap: number;
  maxSupply: number;
  roi: number;
  washTrading: number;
  hypeScore: number;
  adoptionRate: number;
  pdRisk: number;
  icon: string;
  category: string;
  conceptsDescription: string;
}